i18n: i18n for latest projects 🌐
