+++ SharePointFramework/PortfolioWebParts/package.json
@@ -11,7 +11,7 @@
     "watch": "concurrently \"npm run serve\" \"livereload './dist/*.js' -e 'js' -w 250\"",
     "prewatch": "node node_modules/pzl-spfx-tasks --pre-watch --loglevel silent",
     "postwatch": "node node_modules/pzl-spfx-tasks --post-watch --loglevel silent",
-    "serve": "concurrently \"gulp serve-deprecated --locale=nb-no --nobrowser NODE --max-old-space-size=8192\"",
+    "serve": "concurrently \"gulp serve-deprecated --nobrowser NODE --max-old-space-size=8192\"",
     "build": "gulp bundle --ship && gulp package-solution --ship",
     "postversion": "tsc && npm publish",
     "lint": "eslint --ext .ts,.tsx ./src --color --fix --config ../.eslintrc.yaml && npm run prettier",
