+++ SharePointFramework/PortfolioWebParts/src/loc/en-us.js
@@ -143,7 +143,8 @@ define([], function () {
     NoAccessMessage: 'You do not have access to this project',
     NoDefaultViewMessage: 'No default display has been set.',
     NoProjectData: 'Unable to retrieve all data from the project. It may be that you do not have access to the project itself. It may also be that the project has recently been created or that project properties have not been filled in.',
-    NoProjectsFoundMessage: 'No projects found.',
+    NoProjectsFoundMessage: 'There\'s no projects that have been created yet or have shown up in the search.',
+    NoProjectsFoundTitle: 'No projects found',
     NotSet: 'Not set',
     ParentProjectsHeaderText: 'Parent projects',
     ParentProjectsSearchBoxPlaceholderText: 'Search {0} parent projects...',
